Dev back new merge

See merge request s07-webmobile1-sub2/S07P12A602!83

+++ backend/src/main/java/com/jangbo/api/service/FileService.java
@@ -7,9 +7,10 @@ import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.DeleteObjectRequest;
-import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.jangbo.api.request.StoreRegisterPostReq;
+import com.jangbo.db.entity.Store;
+import com.jangbo.db.repository.StoreRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
@@ -20,18 +21,18 @@ import org.springframework.web.server.ResponseStatusException;
 
 import javax.annotation.PostConstruct;
 import java.io.IOException;
-import java.io.InputStream;
 import java.text.SimpleDateFormat;
-import java.util.UUID;
+import java.util.Date;
+
 
 @RequiredArgsConstructor
 @Service
 @Transactional(readOnly = true)
 public class FileService {
-    //private final PostRepository postRepository;
+
+    private final StoreRepository storeRepository;
 
     private AmazonS3 amazonS3;
-
 
     @Value("${cloud.aws.credentials.access-key}")
     private String accessKey;
@@ -55,37 +56,70 @@ public class FileService {
                 .build();
     }
 
+    @Transactional
+    public String uploadImg(StoreRegisterPostReq store, MultipartFile multipartFile) throws IOException {
+        String fileName = createFileName(store.getStoreName(), store.getSellerNo(), multipartFile.getOriginalFilename());
+        amazonS3.putObject(new PutObjectRequest(bucket, fileName, multipartFile.getInputStream(), null)
+                .withCannedAcl(CannedAccessControlList.PublicRead));
+        return fileName;
+    }
 
-    public String fileUpload(StoreRegisterPostReq store, MultipartFile multipartFile) throws IOException{
-        //s3 관련
-        String fileName = createFileName(store.getStoreName(),store.getSellerNo(),multipartFile.getOriginalFilename());
-//        ObjectMetadata objectMetadata = new ObjectMetadata();
-//        objectMetadata.setContentType(multipartFile.getContentType());
-        System.out.println(bucket);
-        System.out.println(fileName);
-        System.out.println(multipartFile.getInputStream());
-            amazonS3.putObject(new PutObjectRequest(bucket, fileName, multipartFile.getInputStream(), null)
-                    .withCannedAcl(CannedAccessControlList.PublicRead));
-        String imgUrl = amazonS3.getUrl(bucket, fileName).toString();
-        System.out.println(imgUrl);
-        return imgUrl;
+    @Transactional
+    public void updateImg(Integer storeNo, MultipartFile file) throws IOException {
+        Store store = storeRepository.findById(storeNo)
+                .orElseThrow(() -> new IllegalAccessError("[storeNo=" + storeNo + "] 해당 상점은 존재하지 않습니다."));
+        Integer sellerNo = store.getSeller().getSellerNo();
+        String fileName = createFileName(store.getStoreName(), sellerNo, file.getOriginalFilename());
+        amazonS3.putObject(new PutObjectRequest(bucket, fileName, file.getInputStream(), null)
+                .withCannedAcl(CannedAccessControlList.PublicRead));
+
+        if(!(store.getStoreImg().equals("default.png"))){ //그 전에 사진이 있었어
+            amazonS3.deleteObject(new DeleteObjectRequest(bucket, store.getStoreImg()));
+            System.out.println(store.getStoreImg()+"지웠어");
+        }
+
+      store.updateImg(fileName);
+        System.out.println(fileName+"생겼어")
+
+
     }
 
     ////////////////////////------------S3관련---------------//////////////////////////////
 
-
-    public void deleteImage(String fileName) {
-        amazonS3.deleteObject(new DeleteObjectRequest(bucket, fileName));
+    @Transactional
+    public Integer deleteImg(Integer storeNo) {
+        Store store = storeRepository.findById(storeNo)
+                .orElseThrow(() -> new IllegalAccessError("[storeNo=" + storeNo + "] 해당 상점은 존재하지 않습니다."));
+        String fileName = store.getStoreImg();
+        if (fileName=="default.png") {
+            return 400;//배드리퀘스트. 삭제할 이미지가 없음
+        } else {
+            amazonS3.deleteObject(new DeleteObjectRequest(bucket, fileName));
+            return 200;
+        }
     }
 
-    private String createFileName(String storeName,int sellerNo,String fileName) {
+    @Transactional
+    public String downloadImg(Integer storeNo) {
+        Store store = storeRepository.findById(storeNo)
+                .orElseThrow(() -> new IllegalAccessError("[storeNo=" + storeNo + "] 해당 상점은 존재하지 않습니다."));
+        if(store.getStoreImg()==null){
+            return amazonS3.getUrl(bucket, "default.png").toString();
+        }else {
+            return amazonS3.getUrl(bucket, store.getStoreImg()).toString();
+        }
+    }
+
+
+    private String createFileName(String storeName, int sellerNo, String fileName) {
+
         SimpleDateFormat dateFormat = new SimpleDateFormat("yyyyMMdd-HHmmSS");
-        String changedName = storeName+"-"+sellerNo+"-"+fileName;
-        System.out.println(changedName);
+        Date date = new Date();
+        String dateForm = dateFormat.format(date);
+        String changedName = storeName + "-" + sellerNo + "-" + dateForm;
         getFileExtension(fileName);
-        return storeName.concat(getFileExtension(fileName));
+        return changedName.concat(getFileExtension(fileName));
     }
-
 
     private String getFileExtension(String fileName) {
         try {
@@ -95,3 +129,6 @@ public class FileService {
         }
     }
 }
+
+
+
